Merged files from Winhost where the latest version wasn't updated in Github.
Winhost and Github should now be in 100% sync.  Please do a DiffMerge
on these files if you are in doubt.  Also, ALWAYS update both Github
and Winhost at the same time as we're adding a couple new contractors
and it will be impossible if these aren't in sync at all times.