Fixed GetPositions must use the store procedure instead of inline SQL. All GetPositions around the site was joined and placed into a new folder called JSON, using the general JSON result format except when data is needed to feed the Autocomplete jQueryUI feature; to return GetPositions in Autocomplete format use the URI
 UrlUtil.JsonPath + "GetPositions/Autocomplete/"

Note that UrlUtil class have now two new properties: JsonPath and
JsonUrl, to link the the JSON folder. Is strongly recommended use
this property, specially from Javascript files.

Now the JSON folder
is under the language folder, but maybe will be moved out (on
the Root folder) because have not localizable content (but have
localizable results - must be do it dynamically), using the
UrlUtil.Json* properties will avoid link problems.

Script.js and Provider-sign-up.js was adapted to the changes.