Communications Updates & LcREST change to payment processing fees
issues #844 & #877

Added BookingCompleted to communications for each booking type and
removed RequestToReview as it’s now included along with a receipt and
payment completed info.

There was a bug in calculating payment processing fees. The values in
bookingType.PaymentProcessingFeePercentage and
bookingtype.FirstTimeBookingFreePercentage were not percentages. Those
were updated but need to be floats and payment processing fee set to
0.029.