Badge components now accept a `src` parameter to retrieve JSON, or an `assertion` parameter to provide assertion data directly.
In the case of `<badge-view/>`, we want to retrieve the badge data remotely from `src`.

In the case of the upcoming `<badge-collection/>`, we're provided all assertion data in a single API call. If `<badge-view/>` can both retrieve this data remotely and accept it as a param, we can reuse it in both places.